refactor benchmark folders structure
